Now transport catalog can response on stat requests (but can't parse it)
The calculation of the shortest path along the sphere and its tests are completed in Sphere namespace.
${SRC_DIRECTORY}/calculateData.h file was removed since it's unused.

Logic of stat requests responses calculation is completed in transportCatalog class.
ParsedRequests now is a structure, which contains two vectors (Stops and Buses). It's turned out that is more convenient for stat requests responses calculation.
MakeBaseRequests::Coordinate structure now is Sphere::Point.