Build refactoring. Now we build protobuf classes as a static library
We need these changes to avoid building bug: we need to build the unit test project first because it runs after building transport_catalog. But we include TransportRouter.h and Router.h to the unit test project to test it. And these files include *pb.h files, generated by protobuf during building of transport_catalog. So when we build the test project, we don't have *pb.h files yet. 
To avoid this problem, we need to run protobuf_generate_cpp(PROTO_SRCS PROTO_HDRS ${PROJECT_PROTOS}) before starting building the unit test project. 
In the current structure of CMakeLists.txt files, it would be inconsistent to place protobuf classes building steps to the top-level CMakeLists.txt. So we need a separated CMakeLists.txt file for protobuf classes building. In this case, the most logical and clear way to do it (from my side of view) is to build protobuf classes as a static library as a first target in the project building process.